refactor(metricsTable): add custom hook to fetch metrics and improve readability
