+++ src/types/Metric.ts
@@ -1,5 +1,4 @@
 export interface Metric {
-    category : string;
     totalProducts: number;
     totalValue: number;
     averagePrice: number;
